Abstracted away batching logic into separate class
