+++ appenders/python/pulse/handlers.py
@@ -1,14 +1,10 @@
-import requests
 import logging
-import sys
 from logging.handlers import MemoryHandler
-import json
-from queue import Queue
-import threading
+from .batcher import PulseBatcher
 
 
 # noinspection PyPep8Naming
-class PulseHandler(MemoryHandler):
+class PulseHandler(PulseBatcher, MemoryHandler):
     """
     A class which sends records to the Pulse log collector. Note that records
     are converted to JSON upon posting. You must provide a
@@ -17,8 +13,7 @@ class PulseHandler(MemoryHandler):
 
     def __init__(self, endpoint, capacity=1000, flushLevel=logging.ERROR, threadCount=1):
         """
-        Initializes PulseHandler using the REST API endpoint. By default,
-        no buffering is done.
+        Initializes PulseHandler using the REST API endpoint, buffer capacity, buffer flush level and thread count.
 
         :param endpoint: The REST API endpoint for the Pulse Log Collector
         :type endpoint: str
@@ -30,36 +25,8 @@ class PulseHandler(MemoryHandler):
         :type threadCount: int
         :rtype: PulseHandler
         """
-        super(PulseHandler, self).__init__(capacity, flushLevel)
-        self.endpoint = endpoint
-        self.debug = False
-
-        # Initialize Threading
-        self.thread_count = threadCount
-        self.queue = Queue()
-        self.threads = list()
-        for i in range(self.thread_count):
-            thread = threading.Thread(target=self.__threadWorker)
-            thread.start()
-            self.threads.append(thread)
-
-        # Initialize Logging
-        self.logger = logging.getLogger(__name__)
-        handler = logging.StreamHandler(sys.stdout)
-        fmt = "[%(asctime)s] %(levelname)s [%(name)s.%(funcName)s:%(lineno)d:%(threadName)s] %(message)s"
-        handler.setFormatter(logging.Formatter(fmt))
-        handler.setLevel(logging.INFO)
-        self.logger.addHandler(handler)
-        self.logger.setLevel(logging.INFO)
-        
-    def setDebug(self):
-        """
-        Set debug mode for MetricWriter.
-        """
-        self.debug = True
-        self.logger.setLevel(logging.DEBUG)
-        for handler in self.logger.handlers:
-            handler.setLevel(logging.DEBUG)
+        PulseBatcher.__init__(self, endpoint, capacity, threadCount)
+        MemoryHandler.__init__(self, capacity, flushLevel)
 
     def setFormatter(self, fmt):
         """
@@ -77,79 +44,42 @@ class PulseHandler(MemoryHandler):
 
         self.formatter = fmt
 
-    def emit(self, buffer):
+    def shouldFlush(self, record):
         """
-        Emit the buffered records, depending on the filters added to the
-        handler.
-
-        :param buffer: List of records to send to Pulse.
-        :type buffer: list(logging.LogRecord)
+        Check for buffer full or a record at the flushLevel or higher.
         """
-        # Filter records in buffer
-        self.logger.debug("Filtering records in buffer")
-        filtered_buffer = [
-            record
-            for record in buffer
-            if self.filter(record)
-        ]
-
-        # If there are records left after filtering, post to Pulse Log Collector
-        self.logger.debug("Putting record buffer into queue")
-        if len(filtered_buffer) > 0:
-            self.queue.put(filtered_buffer, block=False)
-
-    def __threadWorker(self):
-        """
-        Method is used for threading API put requests so that they are non-blocking.
-        """
-        while True:
-            buffer = self.queue.get()
-            if buffer is None:
-                self.logger.debug("Terminating thread")
-                break
-            try:
-                self.logger.debug("Posting records to API endpoint")
-                requests.post(self.endpoint,
-                              json.dumps([self.format(record) for record in buffer]),
-                              headers={"Content-type": "application/json"})
-            except requests.exceptions.RequestException:
-                self.logger.error("---Posting Error---")
-                self.logger.error("---Failed LogRecords Printed Below---")
-                for record in buffer:
-                    self.logger.handle(record)
-            self.queue.task_done()
+        # Note: Calling class method directly to avoid MRO
+        return MemoryHandler.shouldFlush(self, record)
 
     def flush(self):
         """
-        Flush records from buffer and clear buffer.
+        Flush items from buffer and clear buffer.
         """
-        # Emit records in buffer, then clear buffer.
-        self.logger.debug("Flushing log records")
-        self.emit(self.buffer)
-        self.buffer.clear()
+        # Format records before flushing. This couldn't be done sooner because of the LogLevel check
+        # in the shouldFlush method.
+        self.logger.debug("Applying format to log record")
+
+        self.buffer = [
+            self.format(record)
+            for record in self.buffer
+        ]
+
+        PulseBatcher.flush(self)
 
     def handle(self, record):
         """
-        Add record to buffer and flush when appropriate.
+        Add item to buffer and flush when appropriate.
 
         :param record: Record to send to Pulse
         :type record: logging.LogRecord
         """
-        # Append record to buffer and if buffer is at capacity and flush.
-        #
-        # Note: This is the entry point from logging.Logger
-        self.logger.debug("Buffering log record")
-        self.buffer.append(record)
-        if self.shouldFlush(record):
-            self.flush()
+        self.logger.debug("Applying filter to log record")
+        if self.filter(record):
+            super(PulseHandler, self).handle(record)
 
     def close(self):
         """
         Flush remaining records and terminate all threads
         """
-        self.logger.debug("Cleaning up class")
-        super(PulseHandler, self).close()
-        for i in range(self.thread_count):
-            self.queue.put(None)
-        for t in self.threads:
-            t.join()
+        MemoryHandler.close(self)
+        PulseBatcher.close(self)
