+++ appenders/python/pulse/metrics.py
@@ -18,19 +18,14 @@
 # Including Apache 2 license because unix_time_micros method was pulled
 # from Apache Kudu project.
 
-import requests
-import json
 from datetime import datetime
 from pytz import utc
 import six
-import sys
-from queue import Queue
-import threading
-import logging
+from .batcher import PulseBatcher
 
 
 # noinspection PyPep8Naming
-class MetricWriter:
+class MetricWriter(PulseBatcher):
     """
     A class which sends metrics to the Pulse Log Collector.
     """
@@ -50,37 +45,7 @@ class MetricWriter:
         :type threadCount: int
         :rtype: MetricWriter
         """
-        self.endpoint = endpoint
-        self.buffer_capacity = capacity
-        self.buffer = list()
-        self.debug = False
-
-        # Initialize Threading
-        self.thread_count = threadCount
-        self.queue = Queue()
-        self.threads = list()
-        for i in range(self.thread_count):
-            thread = threading.Thread(target=self.__threadWorker)
-            thread.start()
-            self.threads.append(thread)
-
-        # Initialize Logging
-        self.logger = logging.getLogger(__name__)
-        handler = logging.StreamHandler(sys.stdout)
-        fmt = "[%(asctime)s] %(levelname)s [%(name)s.%(funcName)s:%(lineno)d:%(threadName)s] %(message)s"
-        handler.setFormatter(logging.Formatter(fmt))
-        handler.setLevel(logging.INFO)
-        self.logger.addHandler(handler)
-        self.logger.setLevel(logging.INFO)
-
-    def setDebug(self):
-        """
-        Set debug mode for MetricWriter.
-        """
-        self.debug = True
-        self.logger.setLevel(logging.DEBUG)
-        for handler in self.logger.handlers:
-            handler.setLevel(logging.DEBUG)
+        PulseBatcher.__init__(self, endpoint, capacity, threadCount)
 
     def gaugeTimestamp(self, tag, value, timestamp, frmt="%Y-%m-%dT%H:%M:%S.%f"):
         """
@@ -116,70 +81,6 @@ class MetricWriter:
         :rtype: requests.Response
         """
         self.gaugeTimestamp(tag, value, datetime.utcnow())
-
-    def __threadWorker(self):
-        """
-        Method is used for threading API put requests so that they are non-blocking.
-        """
-        while True:
-            buffer = self.queue.get()
-            if buffer is None:
-                self.logger.debug("Terminating thread")
-                break
-            try:
-                self.logger.debug("Posting metrics to API endpoint")
-                self.logger.debug(json.dumps(buffer))
-                resp = requests.post(self.endpoint,
-                                     json.dumps(buffer),
-                                     headers={"Content-type": "application/json"})
-                self.logger.debug("Status: [%s] %s" % (resp.status_code, resp.text))
-            except requests.exceptions.RequestException:
-                self.logger.error("---Posting Error---")
-                self.logger.error("---Failed Metrics Printed Below---")
-                for metric in buffer:
-                    self.logger.info(json.dumps(metric))
-            self.logger.debug("Queue task completed")
-            self.queue.task_done()
-
-    def shouldFlush(self):
-        """
-        Check if the buffer is full.
-
-        :rtype: bool
-        """
-        return len(self.buffer) >= self.buffer_capacity
-
-    def flush(self):
-        """
-        Flush metrics from buffer and clear buffer.
-        """
-        self.logger.debug("Flushing metrics from buffer")
-        self.queue.put(list(self.buffer), block=False)
-        self.buffer.clear()
-
-    def handle(self, metric):
-        """
-        Add metric to buffer and flush when appropriate.
-
-        :param metric: Metric to send to Pulse
-        :type metric: dict
-        """
-        self.logger.debug("Buffering metric")
-        # Append metric to buffer and if buffer is at capacity and flush.
-        self.buffer.append(metric)
-        if self.shouldFlush():
-            self.flush()
-
-    def close(self):
-        """
-        Flush remaining metrics and terminate all threads
-        """
-        self.logger.debug("Cleaning up class")
-        self.flush()
-        for i in range(self.thread_count):
-            self.queue.put(None)
-        for t in self.threads:
-            t.join()
 
     @staticmethod
     def unixTimeMicros(timestamp, frmt="%Y-%m-%dT%H:%M:%S.%f"):
